--- .vscode/settings.json
@@ -1,6 +1,0 @@
-{
-    "java.debug.settings.onBuildFailureProceed": true,
-    "githubPullRequests.ignoredPullRequestBranches": [
-        "main"
-    ]
-}
