+++ Alunos/Thalia Crevilatti/Aula5/Paradigmas.md
